# docs/swagger.yaml
basePath: /api/v1
definitions:
  health.Ping:
    properties:
      response:
        example: hello
        type: string
    type: object
  model.AllPortfoliosViewModel:
    properties:
      portfolios:
        items:
          $ref: '#/definitions/model.PortfolioViewModel'
        type: array
    type: object
  model.Error:
    properties:
      message:
        type: string
    type: object
  model.MetadataViewModel:
    properties:
      createTime:
        example: 02/01/2020 11:12:00
        type: string
      id:
        example: "123884"
        type: string
      lastUpdated:
        example: 02/01/2020 11:12:00
        type: string
    type: object
  model.PortfolioCreateModel:
    properties:
      stocks:
        items:
          $ref: '#/definitions/model.StockViewModel'
        type: array
    type: object
  model.PortfolioID:
    properties:
      id:
        type: string
    type: object
  model.PortfolioViewModel:
    properties:
      metadata:
        $ref: '#/definitions/model.MetadataViewModel'
        type: object
      stocks:
        additionalProperties:
          $ref: '#/definitions/model.StockViewModel'
        type: object
    type: object
  model.StockViewModel:
    properties:
      currency:
        example: CAD
        type: string
      currentPrice:
        example: 105
        type: number
      name:
        example: Canadian Pacific Railway Limited
        type: string
      purchaseDate:
        example: 02/03/2020
        type: string
      purchasePrice:
        example: 10000
        type: number
      quantity:
        example: 100
        type: integer
      ticker:
        example: CP.TO
        type: string
    type: object
info:
  contact: {}
  description: Stores and fetches user and model data
  license:
    name: MIT License
  title: Portfolio Service API
  version: "1.0"
paths:
  /health/hello:
    get:
      description: Non-authenticated endpoint that returns 200 with hello message.
        Used to validate that the service is responsive.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/health.Ping'
      summary: Hello sanity endpoint
      tags:
      - health
  /portfolio:
    get:
      description: Non-authenticated endpoint that returns array of all stored portfolios.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.AllPortfoliosViewModel'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get portfolios endpoint
      tags:
      - portfolio
    post:
      consumes:
      - application/json
      description: Insert portfolio. Returns the portfolio ID.
      parameters:
      - description: Add account
        in: body
        name: portfolio
        required: true
        schema:
          $ref: '#/definitions/model.PortfolioCreateModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PortfolioID'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Creates portfolio a unique ID
      tags:
      - portfolio
  /portfolio/{id}:
    get:
      description: Non-authenticated endpoint that returns a portfolio with matching
        key.
      parameters:
      - description: Portfolio ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PortfolioViewModel'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Error'
      summary: Get portfolios endpoint
      tags:
      - portfolio
swagger: "2.0"